Update of atmos_test branch

+++ UnityProject/Assets/Scripts/Health/Living/SimpleAnimal/SimpleAnimal.cs
@@ -1,5 +1,4 @@
 ﻿using System.Collections;
-using Sprites;
 using UnityEngine;
 using UnityEngine.Networking;
 
@@ -32,7 +31,7 @@ public class SimpleAnimal : HealthBehaviour
 		SetAliveState(deadState);
 	}
 
-	public override int ReceiveAndCalculateDamage(GameObject damagedBy, int damage, DamageType damageType,
+	public override float ReceiveAndCalculateDamage(GameObject damagedBy, float damage, DamageType damageType,
 		BodyPartType bodyPartAim)
 	{
 		base.ReceiveAndCalculateDamage(damagedBy, damage, damageType, bodyPartAim);
